Modified style.css
Moved the para class with a font-size of 22px above the small-para class with a font-size of 14 px in order to give the small-para precedence over the para class. Both classes have the same specificity, but the rules that are defined later in the linked stylesheets or <style> tags win, in this case, after the move, the small para class wins.